Update labels in CFEP 5 to align with survey results
The original CFEP 5 had the labels as rc_package and dev_package. During
the discussionof CFEP 5 (https://github.com/conda-forge/cfep/pull/3), a
discussion and survey ensued about whether the prerelease indicators
should be before the package name or after:

https://github.com/conda-forge/cfep/pull/3#issuecomment-545011203

The results were that the prerelease indicators should be after, but the
text of the CFEP was never updated. This commit updates the CFEP.